requirements.txt
Django==5.0
djangorestframework
pyyaml
requests
django-cors-headers


